backend/requirements.txt
# FastAPI Backend Dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# ML Models
transformers
torch
torchvision
Pillow
numpy
scikit-learn

# Image Processing
opencv-python
grad-cam

# Audio Processing
librosa
soundfile

# Database
aiosqlite

# HTTP Requests
httpx
requests

# Utilities
python-dotenv
langdetect
sentencepiece
protobuf

# AI APIs (Free tier)
google-generativeai
groq